+++ dom/src/main/scala/org/scalajs/dom/FilePropertyBag.scala
@@ -4,5 +4,5 @@ import scala.scalajs.js
 
 /** An options object containing optional attributes for the file. */
 trait FilePropertyBag extends BlobPropertyBag {
-  var lastModified: js.UndefOr[Int] = js.undefined
+  var lastModified: js.UndefOr[Double] = js.undefined
 }
